Skip fbturbo install on non-graphical image
If the image does not have the X server installed, the image
creation will fail in the fbturbo install script.

This commit checks for the xorg/drivers directory before trying
to copy the driver.  If it doesn't exist, then the driver isn't
needed and the copy is skipped.

+++ scripts/root/30_fbturbo.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 [[ -e "$RESOURCEDIR/fbturbo_drv.so" ]] && [[ -e "$RESOURCEDIR/xorg.conf" ]] || exit 0
+[[ -e "$$MNTDIR/usr/lib/xorg/modules/drivers" ]] || exit 0
 echo "Copying X11 fbturbo driver..."
 sudo cp -f "$RESOURCEDIR/fbturbo_drv.so" "$MNTDIR/usr/lib/xorg/modules/drivers" || exit 1
 sudo cp -f "$RESOURCEDIR/xorg.conf" "$MNTDIR/etc/X11" || exit 1
